up model benchmarkgrepl-nonlinear-elliptic with multi level with SER :
The actual design rebuild the model for each level, so the mesh, space, ... are also rebuilt.
This is expansive and also lead to performance issue in eim offline because some expressions mix element of space defined at several level -> but the mesh are rebuilt, feelpp evaluate that is not the same mesh and interpolate.
I do a quick fix (not really nice) by using static mesh objet.