Update qs_stokes.cpp
I was using this template and noticed the results were not correct when changing the viscosity. Also I am not sure to understand why there should be the symmetric gradient in place of the gradient of the trial function. I think it creates problems when imposing the zero-divergence as a constraint.