fix test_integration_ifaces [ci skip]
random failure can appears with this test, depend on the mesh generation with one BOOST_CHECK_CLOSE.
In fact, the result can be a value null or non null. But in the case of a null result, BOOST_CHECK_CLOSE apply a division by 0 and generate an error.
I replace BOOST_CHECK_CLOSE by BOOST_CHECK_SMALL