major changes, insertion of inheritance diamond :
Now, ModelPhysics and ModelAlgebraic inherit of ModelBase. The idea is to have a continous view (ModelPhysics) and discrete view (ModelNumeric).
The modelBase class store the worldComm, prefix, directory, ... some basic informations that these differents views should share.
The only bad side that I see is that we need to call the ModelBase constructor many location.