remove use of deprecated sprintf #2087
Refactor ExporterEnsight to use fmt library and remove use of deprecated sprintf #2087.

This commit replaces calls to the outdated and insecure sprintf function with the modern fmt library. This change improves the overall security of the program, ensuring that we don't leave our application open to buffer overflows. Additionally, this change should improve the speed and readability of the code.