up add support for specx #2009
feat: add support for specx #2009

This commit adds support for building feelpp with specx, allowing users to use statistical performance models to schedule the computation of their problems. The specx CMakePresets.json configuration file is also included in this commit.

Additionally, the file benchmarks/feelpp/99-specx.cpp is added, which showcases an example implementation that computes the area of a Simplex mesh. The implementation utilizes the SpRuntime library from specx and custom `divide_work` function to parallelize the area computation. The benchmarks/feelpp/99-taskflow.cpp file has been removed as it is no longer needed.

/cc @berenger-eu @vincentchabannes @lemoinep