enable mesh distance metric with respect to arbitrary markers #2095
Enable mesh distance metric with respect to arbitrary markers.

This change allows the mesh to support distance metrics based on arbitrary markers using the "gradedfromls" function, which can take a list of markers and various parameters. In order to do this, we have added code to parse the metric expression and set the metric accordingly. This change improves the flexibility of the mesh implementation by allowing it to handle metrics based on any user-specified set of markers.